pyproject.toml
[tool.poetry]
name = "folder2pdf"
version = "0.1.0"
description = ""
authors = ["Daniel Winkler <dan@danwink.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pdfrw = "*"
plumbum = "*"
Pillow = "*"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pylint = "*"
black = "^20.8b1"
pyinstaller = "^4.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
